requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
